+++ src/hid_comm.cpp
@@ -29,12 +29,12 @@ void communication::sendSample(const std::array<float, 15> &game_sample, std::ar
     communication::sendData(tx_data);
 }
 
-void communication::packRawSample(unsigned long timestamp, int deviation, const std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data)
+void communication::packRawSample(uint32_t timestamp, int16_t deviation, const std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data)
 {
     std::array<uint8_t, 4> long_container;
     std::array<uint8_t, 2> int_container;
-    packing::num2bigendbytes<unsigned long>(timestamp, long_container);
-    packing::num2bigendbytes<short>(deviation, int_container);
+    packing::num2bigendbytes<uint32_t>(timestamp, long_container);
+    packing::num2bigendbytes<int16_t>(deviation, int_container);
     std::copy_n(long_container.begin(), long_container.size(), tx_data.begin());
     std::copy_n(int_container.begin(), int_container.size(), tx_data.begin() + 4);
     unsigned int count = 6;
@@ -47,7 +47,7 @@ void communication::packRawSample(unsigned long timestamp, int deviation, const 
     // should have 46 bytes filled by now
 }
 
-void communication::sendSample(unsigned long timestamp, int deviation, std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data)
+void communication::sendSample(uint32_t timestamp, int16_t deviation, std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data)
 {
     communication::packRawSample(timestamp, deviation, raw_sample, tx_data);
     communication::sendData(tx_data);
